@@ -1,20 +1,19 @@
-﻿using Swashbuckle.AspNetCore.Swagger;
+﻿using Microsoft.AspNetCore.Routing.Constraints;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System.Linq;
-using System.Web.Http.Routing.Constraints;
 
 namespace Marsman.ReallySimpleDocumentation
 {
     public class GuidParameterFilter : IParameterFilter
     {
-        public void Apply(IParameter parameter, ParameterFilterContext context)
+        public void Apply(OpenApiParameter parameter, ParameterFilterContext context)
         {
             if (context.PropertyInfo == null &&
-                parameter is NonBodyParameter nonBodyParam &&
                 context.ApiParameterDescription.RouteInfo?.Constraints != null &&
                 context.ApiParameterDescription.RouteInfo.Constraints.Any(c => c is GuidRouteConstraint))
             {
-                nonBodyParam.Format = "uuid";
+                parameter.Schema.Format = "uuid";
             }
         }
     }
