@@ -1,6 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
-using Swashbuckle.AspNetCore.Swagger;
+﻿using Microsoft.Extensions.Options;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System;
 using System.Collections.Generic;
@@ -17,10 +16,24 @@ namespace Marsman.ReallySimpleDocumentation
             this.options = options?.Value ?? throw new ArgumentNullException(nameof(options));
         }
 
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
-            operation.Security = new List<IDictionary<string, IEnumerable<string>>> {
-                    new Dictionary<string, IEnumerable<string>> {{"oauth2", options.Scopes.Keys.ToArray()} }
+            operation.Security = new List<OpenApiSecurityRequirement>
+            {
+                new OpenApiSecurityRequirement
+                {
+                    {
+                        new OpenApiSecurityScheme
+                        {
+                            Reference = new OpenApiReference
+                            {
+                                Type = ReferenceType.SecurityScheme,
+                                Id = "oauth2"
+                            }
+                        },
+                        options.Scopes.Keys.ToList()
+                    }
+                }
             };
         }
     }
