@@ -1,4 +1,4 @@
-﻿using Swashbuckle.AspNetCore.Swagger;
+﻿using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System.Linq;
 using System.Reflection;
@@ -7,14 +7,14 @@ namespace Marsman.ReallySimpleDocumentation
 {
     public class SwaggerIgnoreFilter : ISchemaFilter
     {
-        public void Apply(Schema schema, SchemaFilterContext context)
+        public void Apply(OpenApiSchema schema, SchemaFilterContext context)
         {
             if (schema?.Properties == null)
             {
                 return;
             }
 
-            var excludedProperties = context.SystemType.GetProperties(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance).Where(t => t.GetCustomAttribute<SwaggerIgnoreAttribute>() != null);
+            var excludedProperties = context.Type.GetProperties(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance).Where(t => t.GetCustomAttribute<SwaggerIgnoreAttribute>() != null);
             foreach (var excludedProperty in excludedProperties)
             {
                 if (schema.Properties.ContainsKey(excludedProperty.Name))
