@@ -1,4 +1,4 @@
-﻿using Swashbuckle.AspNetCore.Swagger;
+﻿using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System;
 using System.Collections.Generic;
@@ -9,14 +9,14 @@ namespace Marsman.ReallySimpleDocumentation
 {
     public class SwaggerRequiredFilter : ISchemaFilter
     {
-        public void Apply(Schema schema, SchemaFilterContext context)
+        public void Apply(OpenApiSchema schema, SchemaFilterContext context)
         {
             if (schema?.Properties == null)
             {
                 return;
             }
 
-            var excludedProperties = context.SystemType
+            var excludedProperties = context.Type
                                             .GetProperties(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
                                             .Select(t => (Prop: t, Attr: t.GetCustomAttribute<SwaggerRequiredAttribute>()))
                                             .Where(x => x.Item2 != null);
@@ -29,7 +29,7 @@ namespace Marsman.ReallySimpleDocumentation
 
                     if (schema.Required == null)
                     {
-                        schema.Required = new List<string>();
+                        schema.Required = new HashSet<string>();
                     }
                     schema.Required.Add(propName);
                 }
