@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace Marsman.ReallySimpleDocumentation
@@ -18,7 +18,7 @@ namespace Marsman.ReallySimpleDocumentation
             this.docOptions = docOptions?.Value;
         }
 
-        public void Apply(SwaggerDocument swaggerDoc, DocumentFilterContext context)
+        public void Apply(OpenApiDocument swaggerDoc, DocumentFilterContext context)
         {
             var pathString = hcx.HttpContext.Request.Path;
             if (pathString.Value.Contains("/swaggerui/"))
