@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc.Controllers;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using System.Linq;
 
@@ -7,12 +7,12 @@ namespace Marsman.ReallySimpleDocumentation
 {
     public class UniqueOperationNamesFilter : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             var descriptor = context.ApiDescription.ActionDescriptor;
             if (descriptor is ControllerActionDescriptor cad)
             {
-                operation.OperationId = $"{cad.ControllerName}.{operation.OperationId}";
+                operation.OperationId = $"{cad.ControllerName}.{cad.ActionName}";
                 if (cad.ControllerTypeInfo.GetMethods().Count(x => x.Name == cad.MethodInfo.Name) > 1 && cad.MethodInfo.GetParameters().Length > 0)
                 {
                     operation.OperationId += $".{string.Join(".", cad.MethodInfo.GetParameters().Select(x => x.Name))}";
